# specs/openapi.yaml
openapi: 3.1.0
info:
  title: KINSHASA MOSOLO — API du socle (v1)
  version: 1.0.0
  summary: Plateforme souveraine des recettes de la Ville-Province de Kinshasa
  description: |
    Contrat de référence : `specs/contrat-api.md`. Montants au format `MoneyJSON` (`{"amount": "150.00", "currency": "USD"}`),
    jamais de nombre flottant. Erreurs au format RFC 9457 (`application/problem+json`) avec un `code` stable.

    **Authentification de démonstration** : en-tête `x-demo-user: <userId>` (voir `GET /v1/demo/users`).
    En production : OIDC + clés d'accès (passkeys), mTLS + signatures pour les prestataires.

    Routes marquées `x-extension: true` : compléments du socle, non listés dans le contrat v1 (le frontend peut les ignorer).
  license:
    name: Propriété de la Ville-Province de Kinshasa
servers:
  - url: http://localhost:8080
    description: Développement local (`npm run dev -w backend`)
...
security:
  - demoUser: []
paths:
...
  /v1/legal-rules:
    get:
      tags: [Registre juridique]
      summary: Liste des fiches de règles (tous agents)
      responses:
        '200':
          description: Fiches
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RuleSheet' }
        '403': { $ref: '#/components/responses/Problem' }
    post:
      tags: [Registre juridique]
      summary: Création d'une fiche (R13) — statut BROUILLON
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RuleInput' }
      responses:
        '201':
          description: Fiche créée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RuleSheet' }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/legal-rules/{id}:
    get:
      tags: [Registre juridique]
      summary: Détail d'une fiche avec les entrées requises par la formule
      x-extension: true
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: Fiche
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/RuleSheet' }
                  - type: object
                    properties:
                      requiredInputs: { type: array, items: { type: string } }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/legal-rules/{id}/approve:
    post:
      tags: [Registre juridique]
      summary: Visa (REDACTEUR R13 → VERIFICATEUR_JURIDIQUE R14 → VALIDATEUR_FINANCIER R15 → AUTORITE_PUBLICATION R16)
      description: |
        Quatre personnes distinctes (sinon 403 `SEPARATION_OF_DUTIES`). La 4e approbation publie (`PUBLIEE`),
        puis `ACTIVE` dès la date d'effet. Publication bloquée (422) si un instrument est abrogé
        (`ABROGATED_INSTRUMENT`), non en vigueur (`INSTRUMENT_NOT_IN_FORCE`) ou si la source n'est pas certifiée (`SOURCE_NOT_CERTIFIED`).
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              additionalProperties: false
              properties:
                role: { $ref: '#/components/schemas/ApprovalRole' }
      responses:
        '200':
          description: Fiche après visa
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RuleSheet' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
  /v1/communications/overview:
    get:
      tags: [Communications]
      summary: Synthèse, couverture par canal, dernières délivrances (agents)
      responses:
        '200':
          description: Synthèse
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommunicationsOverview' }
        '403': { $ref: '#/components/responses/Problem' }
  /v1/communications/events:
    get:
      tags: [Communications]
      summary: Catalogue des événements (agents)
      parameters:
        - { name: category, in: query, required: false, schema: { type: string } }
        - { name: mandatory, in: query, required: false, schema: { type: string, enum: ['true', 'false'] } }
      responses:
        '200':
          description: Catalogue
          content:
            application/json:
              schema:
                type: object
                required: [total, categories, events]
                properties:
                  total: { type: integer }
                  categories: { type: array, items: { $ref: '#/components/schemas/EventCategory' } }
                  events: { type: array, items: { $ref: '#/components/schemas/CommunicationEvent' } }
  /v1/communications/preview/{eventCode}:
    get:
      tags: [Communications]
      summary: Aperçu du courriel HTML (charte Ville de Kinshasa + bloc de l'entité émettrice)
      description: |
        Logo `/logo-ville-de-kinshasa.png`, bandeau tricolore, bloc entité (DGIPK, DGTK, MINFIN…), mention d'avis obligatoire,
        « La version française fait foi. » si `lang ≠ fr`, pied « Plateforme KINSHASA MOSOLO — réalisée par Groupe Nseya ».
        L'objet (encodé URI) est aussi renvoyé dans l'en-tête `x-mosolo-subject`.
      parameters:
        - { name: eventCode, in: path, required: true, schema: { type: string, examples: [assessment.issued] } }
        - { $ref: '#/components/parameters/Lang' }
        - { name: entity, in: query, required: false, schema: { $ref: '#/components/schemas/EntityCode' } }
      responses:
        '200':
          description: Courriel rendu
          headers:
            x-mosolo-subject: { schema: { type: string } }
            content-language: { schema: { type: string } }
          content:
            text/html:
              schema: { type: string }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/communications/test:
    post:
      tags: [Communications]
      summary: Envoi de test à soi-même (statut `journalise` en bac à sable si pas de clé fournisseur)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventCode]
              additionalProperties: false
              properties:
                eventCode: { type: string }
                entity: { $ref: '#/components/schemas/EntityCode' }
                lang: { $ref: '#/components/schemas/LanguageCode' }
      responses:
        '201':
          description: Délivrances
          content:
            application/json:
              schema:
                type: object
                required: [eventCode, entity, sandbox, deliveries]
                properties:
                  eventCode: { type: string }
                  entity: { type: string }
                  sandbox: { type: boolean }
                  deliveries: { type: array, items: { $ref: '#/components/schemas/Delivery' } }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    demoUser:
      type: apiKey
      in: header
      name: x-demo-user
      description: Identifiant d'un utilisateur de démonstration (voir GET /v1/demo/users). DÉMO UNIQUEMENT.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    Lang:
      name: lang
      in: query
      required: false
      description: Sinon `Accept-Language` ; français par défaut (langue de référence).
      schema: { $ref: '#/components/schemas/LanguageCode' }
  responses:
    Problem:
      description: Erreur RFC 9457
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status, detail, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          description: Code stable
          examples: [RULE_NOT_EXECUTABLE, FORBIDDEN, SEPARATION_OF_DUTIES, IDEMPOTENCY_KEY_REUSED, NONCE_REPLAYED, FX_RATE_MISSING, PROVIDER_UNAVAILABLE, AMOUNT_NOT_REPRESENTABLE, PROVIDER_ORDER_MISMATCH]
        instance: { type: string }
      additionalProperties: true
    CurrencyCode:
      type: string
      enum: [CDF, USD, EUR, GBP, CAD, CHF, ZAR, XAF, AOA, ZMW, RWF, UGX, KES, TZS, BIF, CNY, AED]
    LanguageCode:
      type: string
      enum: [fr, ln, sw, kg, lua, en]
    EntityCode:
      type: string
      enum: [GOUVERNORAT, MINFIN, DGIPK, DGTK, TRESOR, AUDIT, PLATEFORME, CONTENTIEUX, PUBLIC]
...
    ApprovalRole:
      type: string
      enum: [REDACTEUR, VERIFICATEUR_JURIDIQUE, VALIDATEUR_FINANCIER, AUTORITE_PUBLICATION]
    RuleStatus:
      type: string
      enum: [A_VERIFIER, BROUILLON, REVUE_JURIDIQUE, REVUE_FINANCIERE, APPROUVEE, PUBLIEE, ACTIVE, SUSPENDUE, EXPIREE, ABROGEE, ARCHIVEE]
    RuleInput:
      type: object
      additionalProperties: false
      required: [code, revenueCategory, label, legalInstrumentIds, articles, competentAuthority, administeringEntity, taxableEvent, liableParty, baseDefinition, formula, rateTable, currency, rounding, periodicity, dueRule, effectiveFrom, beneficiaryAccountAlias, appealPath, sourceVerification]
      properties:
        code: { type: string, pattern: '^[A-Z0-9-]{2,40}$' }
        revenueCategory: { type: string, enum: [IMPOT_PROVINCIAL, INTERET_COMMUN, PROVINCIAL_SPECIFIQUE, RECETTE_ETD, RECETTE_CENTRALE, PARTAGEE, DROIT_ADMINISTRATIF, REDEVANCE_SERVICE, PENALITE, CONCESSION_DOMANIALE, RECETTE_COMMERCIALE, ACTE_REQUIS] }
        label: { type: string }
        legalInstrumentIds: { type: array, items: { type: string }, minItems: 1 }
        articles: { type: array, items: { type: string }, minItems: 1 }
        competentAuthority: { type: string }
        administeringEntity: { type: string }
        taxableEvent: { type: string }
        liableParty: { type: string }
        withholdingAgent: { type: string }
        baseDefinition: { type: string }
        formula:
          type: string
          description: 'Langage de règles : nombres décimaux, identifiants, + - * /, parenthèses, max(), min(). Identifiant résolu dans les entrées, puis `rateTable[nom]`, puis `rateTable["nom:<rang>"]`.'
          examples: ['max(0, loyers_percus * taux / 100 - retenues_imputees)', 'forfait']
        rateTable: { type: object, additionalProperties: { type: string }, examples: [{ 'forfait:1': '450', 'forfait:2': '150' }] }
        currency: { $ref: '#/components/schemas/CurrencyCode' }
        rounding: { type: string, enum: [HALF_UP, HALF_EVEN, DOWN, UP] }
        periodicity: { type: string, enum: [ANNUELLE, MENSUELLE, PONCTUELLE] }
        dueRule: { type: string }
        exemptions: { type: array, items: { type: object, properties: { basis: { type: string }, proof: { type: string } } } }
        penalties: { type: array, items: { type: object, properties: { basis: { type: string }, description: { type: string } } } }
        effectiveFrom: { type: string, format: date }
        effectiveTo: { type: string, format: date }
        beneficiaryAccountAlias: { type: string, description: Alias du coffre (jamais un numéro de compte) }
        appealPath: { type: string }
        sourceVerification: { type: string, enum: [OFFICIEL_CERTIFIE, PRESSE, DOCUMENT_DE_TRAVAIL, AUCUNE] }
        changeReason: { type: string }
    RuleSheet:
      allOf:
        - $ref: '#/components/schemas/RuleInput'
        - type: object
          required: [id, version, status, approvals]
          properties:
            id: { type: string }
            version: { type: integer }
            status: { $ref: '#/components/schemas/RuleStatus' }
            approvals:
              type: array
              items:
                type: object
                properties:
                  role: { $ref: '#/components/schemas/ApprovalRole' }
                  userId: { type: string }
                  at: { type: string, format: date-time }
            supersedesVersionId: { type: string }
            createdBy: { type: string }
            createdAt: { type: string, format: date-time }
            publishedAt: { type: string, format: date-time }
            activatedAt: { type: string, format: date-time }
            sample: { type: boolean, description: Fiche modèle de l'Annexe B (A_VERIFIER) }
            demo: { type: boolean, description: Règle fictive de démonstration }
...
    EventCategory:
      type: object
      properties:
        code: { type: string }
        libelle: { type: string }
    Channel:
      type: string
      enum: [email, in-app, sms, push, whatsapp, ussd, svi, courrier]
    CommunicationEvent:
      type: object
      properties:
        code: { type: string }
        categorie: { type: string }
        libelle: { type: string }
        objet: { type: string }
        gravite: { type: string, enum: [info, success, warning, critical] }
        canaux_defaut: { type: array, items: { $ref: '#/components/schemas/Channel' } }
        whatsapp_optin: { type: boolean }
        obligatoire: { type: boolean }
        public: { type: array, items: { type: string, enum: [C, G, X] } }
    Delivery:
      type: object
      required: [id, at, eventCode, channel, recipientId, status, provider]
      properties:
        id: { type: string }
        at: { type: string, format: date-time }
        eventCode: { type: string }
        category: { type: string }
        channel: { $ref: '#/components/schemas/Channel' }
        recipientId: { type: string }
        recipientKind: { type: string, enum: [taxpayer, user] }
        recipientMasked: { type: string }
        status: { type: string, enum: [en_file, envoye, delivre, lu, echoue, journalise, supprime_par_preference] }
        provider: { type: string }
        providerMode: { type: string, enum: [internal, live, sandbox, aucun] }
        mandatory: { type: boolean }
        entity: { type: string }
        lang: { $ref: '#/components/schemas/LanguageCode' }
        attempts: { type: integer }
        contentHash: { type: string }
    CommunicationsOverview:
      type: object
      required: [catalogue, categories, mandatory, delivered, attempted, channelsWired, coverage, recent]
      properties:
        catalogue: { type: integer, examples: [255] }
        categories: { type: integer, examples: [23] }
        mandatory: { type: integer, examples: [126] }
        delivered: { type: integer }
        attempted: { type: integer }
        sandboxLogged: { type: integer }
        suppressedByPreference: { type: integer }
        channelsWired: { type: integer, description: Canaux raccordés (in-app + canaux avec clé fournisseur) }
        channelsTotal: { type: integer, examples: [8] }
        coverage: { type: object, additionalProperties: { type: integer }, description: Événements diffusés par défaut par canal }
        channels:
          type: array
          items:
            type: object
            properties:
              channel: { $ref: '#/components/schemas/Channel' }
              provider: { type: string }
              mode: { type: string, enum: [internal, live, sandbox] }
              wired: { type: boolean }
              defaultEvents: { type: integer }
              sent: { type: integer }
        recent: { type: array, items: { $ref: '#/components/schemas/Delivery' } }
...